Split tenant Keycloak user projection
